// src/types.ts
/**
 * Longitude and latitude.
 *
 * @beta
 */
export interface LngLat {
  /** Longitude. */
  lng: number;
  /** Latitude. */
  lat: number;
}

/**
 * RGBA components.
 *
 * @beta
 */
export interface RGBA {
  /** Red component. [0, 1]. */
  red: number;
  /** Green component. [0, 1]. */
  green: number;
  /** Blue component. [0, 1]. */
  blue: number;
  /** Alpha component. [0, 1]. */
  alpha: number;
}